refactor: remove unused files and consolidate configuration and logging setup
