--- app/pipelines.py
@@ -1,3 +1,0 @@
-# pipelines.py: Process the data after it’s scraped, such as cleaning or storing it in a database.
-
-# Create MD or TXT files from the scraped data
