+++ app/services/scraper.py
@@ -1,8 +1,75 @@
 import logging
 from typing import List, Dict, Any, Optional
-from app.utils.scraper import WebScraper
+import requests
+import re
+import time
+from scrapy import Selector
+from fake_useragent import UserAgent
+from requests.adapters import HTTPAdapter
+from urllib3.util.retry import Retry
+from app.utils.config import config
 from app.models.scraper import PageSnapshot, CleanedText
 from sqlalchemy.orm import Session
+from datetime import datetime, timezone
+
+class WebScraper:
+    def _setup_session(self) -> requests.Session:
+        session = requests.Session()
+        retries = Retry(
+            total=config.max_retries,
+            backoff_factor=0.3,
+            status_forcelist=[429, 500, 502, 503, 504],
+            allowed_methods=frozenset(["GET", "POST"])
+        )
+        session.mount("https://", HTTPAdapter(max_retries=retries))
+        session.mount("http://", HTTPAdapter(max_retries=retries))
+        return session
+    
+    def _get_user_agent(self) -> str:
+        try:
+            ua = UserAgent()
+            return ua.random
+        except Exception as e:
+            logging.warning(f"fake-useragent failed, using fallback: {e}")
+            return "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
+    
+    def _extract_text(self, html: str) -> str:
+        # Clean HTML
+        cleaned_html = re.sub(r'(?is)<(script|style)[^>]*>.*?</\1>', '', html)
+        cleaned_html = re.sub(r'<!--.*?-->', '', cleaned_html, flags=re.S)
+
+        # Extract text
+        selector = Selector(text=cleaned_html)
+        text = ' '.join(selector.xpath('//body//text()').getall())
+        return re.sub(r'\s+', ' ', text).strip()
+    
+    def scrape_url(self, url: str) -> PageSnapshot:        
+        session = self._setup_session()
+        user_agent = self._get_user_agent()
+        headers = {"User-Agent": user_agent}
+        
+        logging.info(f"Scraping {url}")
+        
+        try:
+            response = session.get(url, headers=headers, timeout=config.timeout)
+            response.raise_for_status()
+        except requests.RequestException as e:
+            logging.error(f"Error scraping {url}: {e}")
+            raise
+        
+        text = self._extract_text(response.text)
+      
+        time.sleep(config.scrape_delay)
+        
+        url_hash = PageSnapshot.compute_hash_from_url(url)
+
+        return PageSnapshot(
+            url=url,
+            user_agent=user_agent,
+            content=text,
+            retrieved_at=datetime.now(timezone.utc),
+            url_hash=url_hash
+        )
 
 class ScraperService:
     def __init__(self, db: Session):
