--- app/settings.py
@@ -1,5 +1,0 @@
-# TODO : Add settings
-
-# FastAPI settings
-
-# Scrapy settings
