--- app/spiders/__init__.py
@@ -1,1 +1,0 @@
-# spiders/: The directory where you’ll create your spiders, the core components of Scrapy that define how to scrape websites.
